create seperate chart.go, repository.go files
