+++ harbor/project.go
@@ -1,9 +1,5 @@
 package harbor
 
-import (
-	"fmt"
-)
-
 type ProjectReq struct {
 	CountLimit  int64  `json:"count_limit,omitempty"`
 	ProjectName string `json:"project_name,omitempty"`
@@ -37,30 +33,6 @@ type ProjectMetadata struct {
 	PreventVul           string `json:"prevent_vul,omitempty"`
 }
 
-type Repository struct {
-	ID          int           `json:"id"`
-	Name        string        `json:"name"`
-	ProjectID   int           `json:"project_id"`
-	Description string        `json:"description"`
-	PullCount   int           `json:"pull_count"`
-	StarCount   int           `json:"star_count"`
-	TagsCount   int           `json:"tags_count"`
-	Labels      []interface{} `json:"labels"`
-	// CreationTime time.Time     `json:"creation_time"`
-	// UpdateTime   time.Time     `json:"update_time"`
-}
-
-type Chart struct {
-	Name          string    `json:"name"`
-	TotalVersions int       `json:"total_versions"`
-	LatestVersion string    `json:"latest_version"`
-	// Created       time.Time `json:"created"`
-	// Updated       time.Time `json:"updated"`
-	Icon          string    `json:"icon"`
-	Home          string    `json:"home"`
-	Deprecated    bool      `json:"deprecated"`
-}
-
 func (client *Client) GetProject(id string) (*Project, error) {
 	var project *Project
 
@@ -83,54 +55,6 @@ func (client *Client) NewProject(project *ProjectReq) (string, error) {
 
 func (client *Client) UpdateProject(id string, project *ProjectReq) error {
 	return client.put(id, project)
-}
-
-func (client *Client) GetRepositories(id string) ([]*Repository, error) {
-	var repositories []*Repository
-	err := client.get(fmt.Sprintf("/repositories?project_id=%s", id), &repositories, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return repositories, nil
-}
-
-func (client *Client) GetCharts(id string) ([]*Chart, error) {
-	var charts []*Chart
-	err := client.get(fmt.Sprintf("/chartrepo/%s/charts", id), &charts, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return charts, nil
-}
-
-func (client *Client) DeleteCharts(project string, charts []*Chart) error {
-	for _, chart := range charts {
-		err := client.DeleteChart(project, chart.Name)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (client *Client) DeleteRepositories(repos []*Repository) error {
-	for _, repo := range repos {
-		err := client.DeleteRepository(repo.Name)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (client *Client) DeleteRepository(id string) error {
-	return client.delete(fmt.Sprintf("/repositories/%s", id), nil)
-}
-
-func (client *Client) DeleteChart(project string, chart string) error {
-	return client.delete(fmt.Sprintf("/chartrepo/%s/charts/%s", project, chart), nil)
 }
 
 func (client *Client) DeleteProject(id string) error {
